更新测试用例和代码
向量问题处理
